Moved Books to its own subrouter
