+++ cmd/main/main.go
@@ -2,22 +2,19 @@ package main
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"log"
 	"net/http"
 	"os"
 	"time"
 
+	"github.com/cadimoduarte/go-rest-api/pkg/main/api/books"
 	"github.com/cadimoduarte/go-rest-api/pkg/main/api/jokes"
 	"github.com/cadimoduarte/go-rest-api/pkg/main/config"
-	"github.com/cadimoduarte/go-rest-api/pkg/main/helper"
-	"github.com/cadimoduarte/go-rest-api/pkg/main/models"
 
 	"github.com/gorilla/mux"
 	"github.com/joho/godotenv"
 
-	"go.mongodb.org/mongo-driver/bson/primitive"
 	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/options"
 )
@@ -65,9 +62,8 @@ func connectOnMongo() *mongo.Client {
 		}
 		log.Println("Trying again...")
 		connectOnMongo()
-	} //else {
-	//app.Resources["mongo"] = client
-	//}
+	}
+
 	return client
 }
 
@@ -79,114 +75,14 @@ func main() {
 	jokesRouter := jokes.Router{}
 	jokesRouter.ConfigRouter(router.PathPrefix("/api/jokes").Subrouter())
 
-	router.HandleFunc("/", homeLink)
+	booksRouter := books.Router{
+		MongoClient: mongoClient,
+	}
+	booksRouter.ConfigRouter(router.PathPrefix("/api/books").Subrouter())
 
-	// arrange our route
-	router.HandleFunc("/api/books", getBooks).Methods("GET")
-	router.HandleFunc("/api/books/{id}", getBook).Methods("GET")
-	router.HandleFunc("/api/books", createBook).Methods("POST")
-	router.HandleFunc("/api/books/{id}", updateBook).Methods("PUT")
-	router.HandleFunc("/api/books/{id}", deleteBook).Methods("DELETE")
+	router.HandleFunc("/", homeLink)
 
 	// set our port address
 	fmt.Println("Starting server")
 	log.Fatal(http.ListenAndServe(":8080", router))
-}
-
-func getBooks(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-
-	var book = models.Book{}
-
-	books, err := book.Load(mongoClient)
-
-	if err != nil {
-		//TODO: create new error handlers
-		helper.GetError(err, w)
-		return
-	}
-
-	json.NewEncoder(w).Encode(books) // encode similar to serialize process.
-}
-
-func getBook(w http.ResponseWriter, r *http.Request) {
-	// set header.
-	w.Header().Set("Content-Type", "application/json")
-
-	var params = mux.Vars(r)
-	id, _ := primitive.ObjectIDFromHex(params["id"])
-
-	var book = models.Book{
-		ID: id,
-	}
-
-	err := book.Get(mongoClient)
-
-	if err != nil {
-		helper.GetError(err, w)
-		return
-	}
-
-	json.NewEncoder(w).Encode(book)
-
-}
-
-func createBook(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-
-	var book models.Book
-
-	_ = json.NewDecoder(r.Body).Decode(&book)
-
-	err := book.Insert(mongoClient)
-
-	if err != nil {
-		helper.GetError(err, w)
-		return
-	}
-
-	json.NewEncoder(w).Encode(book)
-}
-
-func updateBook(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-
-	var params = mux.Vars(r)
-
-	//Get id from parameters
-	id, _ := primitive.ObjectIDFromHex(params["id"])
-
-	book := models.Book{ID: id}
-
-	_ = json.NewDecoder(r.Body).Decode(&book)
-
-	err := book.Update(mongoClient)
-
-	if err != nil {
-		helper.GetError(err, w)
-		return
-	}
-
-	fmt.Println(book)
-
-	json.NewEncoder(w).Encode(book)
-}
-
-func deleteBook(w http.ResponseWriter, r *http.Request) {
-	// Set header
-	w.Header().Set("Content-Type", "application/json")
-
-	// get params
-	var params = mux.Vars(r)
-
-	// string to primitve.ObjectID
-	id, _ := primitive.ObjectIDFromHex(params["id"])
-
-	book := models.Book{ID: id}
-
-	deleteErr := book.Delete(mongoClient)
-
-	if deleteErr != nil {
-		return
-	}
 }
